Blue Tooth Dongle Application
BLE Dongle Application for PC to receive the data from bluetooth dongle
and get it send to web with web service with local database of
refreshing.
Two major application was zipped, one is for blue tooth recognized with
Serial port to send data to web. The other one is for PC as web service
server with local database for refreshing in web browser.